Move resolver controller utility to shared utils
