+++ src/plugin/src/options/line-comment-options.js
@@ -1,5 +1,8 @@
-import { isNonEmptyString, isRegExpLike } from "../shared/index.js";
-import { createResolverController } from "../shared/resolver-controller.js";
+import {
+    isNonEmptyString,
+    isRegExpLike,
+    createResolverController
+} from "../shared/index.js";
 
 const LINE_COMMENT_BANNER_DETECTION_MIN_SLASHES = 5;
 const LINE_COMMENT_BANNER_STANDARD_LENGTH = 60;
